+++ DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/basic/DbT1009ShikakuShutokuJogaishaEntity.java
@@ -11,6 +11,7 @@ import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.biz.LasdecCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import java.util.Objects;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * DbT1009ShikakuShutokuJogaishaの項目定義クラスです
@@ -18,6 +19,7 @@ import java.util.Objects;
  */
 public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT1009ShikakuShutokuJogaishaEntity> implements IDbAccessable {
 // <editor-fold defaultstate="collapsed" desc="Created By POJO Tool ver 1.3.9">
+
     @TableName
     public static final RString TABLE_NAME = new RString("DbT1009ShikakuShutokuJogaisha");
 
@@ -32,7 +34,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
     @PrimaryKey
     private ShikibetsuCode shikibetsuCode;
     @PrimaryKey
-    private int rirekiNo;
+    private Decimal rirekiNo;
     private LasdecCode shichosonCode;
     private RString shikakuShutokuJogaiRiyu;
     private FlexibleDate shikakuShutokuJogaiYMD;
@@ -40,7 +42,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * insertDantaiCdのgetメソッドです。
-     * 
+     *
      * @return insertDantaiCd
      */
     public RString getInsertDantaiCd() {
@@ -49,7 +51,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * insertDantaiCdのsetメソッドです。
-     * 
+     *
      * @param insertDantaiCd insertDantaiCd
      */
     public void setInsertDantaiCd(RString insertDantaiCd) {
@@ -58,7 +60,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * isDeletedのgetメソッドです。
-     * 
+     *
      * @return isDeleted
      */
     public boolean getIsDeleted() {
@@ -67,17 +69,16 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * isDeletedのsetメソッドです。
-     * 
+     *
      * @param isDeleted isDeleted
      */
     public void setIsDeleted(boolean isDeleted) {
         this.isDeleted = isDeleted;
     }
 
-
     /**
      * lastUpdateReamsLoginIdのsetメソッドです。
-     * 
+     *
      * @param lastUpdateReamsLoginId lastUpdateReamsLoginId
      */
     public void setLastUpdateReamsLoginId(RString lastUpdateReamsLoginId) {
@@ -86,7 +87,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 識別コードのgetメソッドです。
-     * 
+     *
      * @return 識別コード
      */
     public ShikibetsuCode getShikibetsuCode() {
@@ -95,7 +96,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 識別コードのsetメソッドです。
-     * 
+     *
      * @param shikibetsuCode 識別コード
      */
     public void setShikibetsuCode(ShikibetsuCode shikibetsuCode) {
@@ -104,25 +105,25 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 履歴番号のgetメソッドです。
-     * 
+     *
      * @return 履歴番号
      */
-    public int getRirekiNo() {
+    public Decimal getRirekiNo() {
         return rirekiNo;
     }
 
     /**
      * 履歴番号のsetメソッドです。
-     * 
+     *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(int rirekiNo) {
+    public void setRirekiNo(Decimal rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
 
     /**
      * 市町村コードのgetメソッドです。
-     * 
+     *
      * @return 市町村コード
      */
     public LasdecCode getShichosonCode() {
@@ -131,7 +132,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 市町村コードのsetメソッドです。
-     * 
+     *
      * @param shichosonCode 市町村コード
      */
     public void setShichosonCode(LasdecCode shichosonCode) {
@@ -140,7 +141,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外理由のgetメソッドです。
-     * 
+     *
      * @return 資格取得除外理由
      */
     public RString getShikakuShutokuJogaiRiyu() {
@@ -149,7 +150,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外理由のsetメソッドです。
-     * 
+     *
      * @param shikakuShutokuJogaiRiyu 資格取得除外理由
      */
     public void setShikakuShutokuJogaiRiyu(RString shikakuShutokuJogaiRiyu) {
@@ -158,7 +159,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外年月日のgetメソッドです。
-     * 
+     *
      * @return 資格取得除外年月日
      */
     public FlexibleDate getShikakuShutokuJogaiYMD() {
@@ -167,7 +168,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外年月日のsetメソッドです。
-     * 
+     *
      * @param shikakuShutokuJogaiYMD 資格取得除外年月日
      */
     public void setShikakuShutokuJogaiYMD(FlexibleDate shikakuShutokuJogaiYMD) {
@@ -176,7 +177,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外解除年月日のgetメソッドです。
-     * 
+     *
      * @return 資格取得除外解除年月日
      */
     public FlexibleDate getShikakuShutokuJogaiKaijoYMD() {
@@ -185,7 +186,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * 資格取得除外解除年月日のsetメソッドです。
-     * 
+     *
      * @param shikakuShutokuJogaiKaijoYMD 資格取得除外解除年月日
      */
     public void setShikakuShutokuJogaiKaijoYMD(FlexibleDate shikakuShutokuJogaiKaijoYMD) {
@@ -194,9 +195,9 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * このエンティティの主キーが他の{@literal DbT1009ShikakuShutokuJogaishaEntity}と等しいか判定します。
-     * 
+     *
      * @param other 比較するエンティティ
-     * @@return 
+     * @@return
      * 比較するエンティティが同じ主キーを持つ{@literal DbT1009ShikakuShutokuJogaishaEntity}の場合{@literal true}、それ以外の場合は{@literal false}
      */
     @Override
@@ -228,6 +229,7 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
 
     /**
      * {@inheritDoc}
+     *
      * @return {@inheritDoc}
      */
     @Override
@@ -236,8 +238,4 @@ public class DbT1009ShikakuShutokuJogaishaEntity extends DbTableEntityBase<DbT10
     }
 
 // </editor-fold>
-
-
-
-
 }
