+++ DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/basic/DbT3006KyotakuKeikakuJigyoshaSakuseiEntity.java
@@ -12,6 +12,7 @@ import jp.co.ndensan.reams.uz.uza.lang.FlexibleYearMonth;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.db.dbz.definition.valueobject.domain.JigyoshaNo;
 import java.util.Objects;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * DbT3006KyotakuKeikakuJigyoshaSakuseiの項目定義クラスです
@@ -19,6 +20,7 @@ import java.util.Objects;
  */
 public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBase<DbT3006KyotakuKeikakuJigyoshaSakuseiEntity> implements IDbAccessable {
 // <editor-fold defaultstate="collapsed" desc="Created By POJO Tool ver 1.4.1">
+
     @TableName
     public static final RString TABLE_NAME = new RString("DbT3006KyotakuKeikakuJigyoshaSakusei");
 
@@ -35,7 +37,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
     @PrimaryKey
     private FlexibleYearMonth taishoYM;
     @PrimaryKey
-    private int rirekiNo;
+    private Decimal rirekiNo;
     private FlexibleDate tekiyoKaishiYMD;
     private FlexibleDate tekiyoShuryoYMD;
     private RString sakuseiKubunCode;
@@ -46,7 +48,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * insertDantaiCdのgetメソッドです。
-     * 
+     *
      * @return insertDantaiCd
      */
     public RString getInsertDantaiCd() {
@@ -55,7 +57,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * insertDantaiCdのsetメソッドです。
-     * 
+     *
      * @param insertDantaiCd insertDantaiCd
      */
     public void setInsertDantaiCd(RString insertDantaiCd) {
@@ -64,7 +66,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * isDeletedのgetメソッドです。
-     * 
+     *
      * @return isDeleted
      */
     public boolean getIsDeleted() {
@@ -73,17 +75,16 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * isDeletedのsetメソッドです。
-     * 
+     *
      * @param isDeleted isDeleted
      */
     public void setIsDeleted(boolean isDeleted) {
         this.isDeleted = isDeleted;
     }
 
-
     /**
      * lastUpdateReamsLoginIdのsetメソッドです。
-     * 
+     *
      * @param lastUpdateReamsLoginId lastUpdateReamsLoginId
      */
     public void setLastUpdateReamsLoginId(RString lastUpdateReamsLoginId) {
@@ -92,7 +93,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 被保険者番号のgetメソッドです。
-     * 
+     *
      * @return 被保険者番号
      */
     public HihokenshaNo getHihokenshano() {
@@ -101,7 +102,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 被保険者番号のsetメソッドです。
-     * 
+     *
      * @param hihokenshano 被保険者番号
      */
     public void setHihokenshano(HihokenshaNo hihokenshano) {
@@ -110,7 +111,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 対象年月のgetメソッドです。
-     * 
+     *
      * @return 対象年月
      */
     public FlexibleYearMonth getTaishoYM() {
@@ -119,7 +120,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 対象年月のsetメソッドです。
-     * 
+     *
      * @param taishoYM 対象年月
      */
     public void setTaishoYM(FlexibleYearMonth taishoYM) {
@@ -128,25 +129,25 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 履歴番号のgetメソッドです。
-     * 
+     *
      * @return 履歴番号
      */
-    public int getRirekiNo() {
+    public Decimal getRirekiNo() {
         return rirekiNo;
     }
 
     /**
      * 履歴番号のsetメソッドです。
-     * 
+     *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(int rirekiNo) {
+    public void setRirekiNo(Decimal rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
 
     /**
      * 適用開始年月日のgetメソッドです。
-     * 
+     *
      * @return 適用開始年月日
      */
     public FlexibleDate getTekiyoKaishiYMD() {
@@ -155,7 +156,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 適用開始年月日のsetメソッドです。
-     * 
+     *
      * @param tekiyoKaishiYMD 適用開始年月日
      */
     public void setTekiyoKaishiYMD(FlexibleDate tekiyoKaishiYMD) {
@@ -164,7 +165,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 適用終了年月日のgetメソッドです。
-     * 
+     *
      * @return 適用終了年月日
      */
     public FlexibleDate getTekiyoShuryoYMD() {
@@ -173,7 +174,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 適用終了年月日のsetメソッドです。
-     * 
+     *
      * @param tekiyoShuryoYMD 適用終了年月日
      */
     public void setTekiyoShuryoYMD(FlexibleDate tekiyoShuryoYMD) {
@@ -184,7 +185,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
      * 作成区分コードのgetメソッドです。
      * <br/>
      * <br/>1：指定居宅介護支援事業者作成、2：基準該当居宅介護支援事業者作成、4：介護予防支援事業者作成
-     * 
+     *
      * @return 作成区分コード
      */
     public RString getSakuseiKubunCode() {
@@ -195,7 +196,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
      * 作成区分コードのsetメソッドです。
      * <br/>
      * <br/>1：指定居宅介護支援事業者作成、2：基準該当居宅介護支援事業者作成、4：介護予防支援事業者作成
-     * 
+     *
      * @param sakuseiKubunCode 作成区分コード
      */
     public void setSakuseiKubunCode(RString sakuseiKubunCode) {
@@ -204,7 +205,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 計画事業者番号のgetメソッドです。
-     * 
+     *
      * @return 計画事業者番号
      */
     public JigyoshaNo getKeikakuJigyoshaNo() {
@@ -213,7 +214,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 計画事業者番号のsetメソッドです。
-     * 
+     *
      * @param keikakuJigyoshaNo 計画事業者番号
      */
     public void setKeikakuJigyoshaNo(JigyoshaNo keikakuJigyoshaNo) {
@@ -224,7 +225,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
      * 委託先事業者番号のgetメソッドです。
      * <br/>
      * <br/>地域包括支援センターの委託先
-     * 
+     *
      * @return 委託先事業者番号
      */
     public JigyoshaNo getItakusakiJigyoshaNo() {
@@ -235,7 +236,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
      * 委託先事業者番号のsetメソッドです。
      * <br/>
      * <br/>地域包括支援センターの委託先
-     * 
+     *
      * @param itakusakiJigyoshaNo 委託先事業者番号
      */
     public void setItakusakiJigyoshaNo(JigyoshaNo itakusakiJigyoshaNo) {
@@ -244,7 +245,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 事業者変更年月日のgetメソッドです。
-     * 
+     *
      * @return 事業者変更年月日
      */
     public FlexibleDate getJigyoshaHenkoYMD() {
@@ -253,7 +254,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 事業者変更年月日のsetメソッドです。
-     * 
+     *
      * @param jigyoshaHenkoYMD 事業者変更年月日
      */
     public void setJigyoshaHenkoYMD(FlexibleDate jigyoshaHenkoYMD) {
@@ -262,7 +263,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 事業者変更事由のgetメソッドです。
-     * 
+     *
      * @return 事業者変更事由
      */
     public RString getJigyoshaHenkoJiyu() {
@@ -271,7 +272,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * 事業者変更事由のsetメソッドです。
-     * 
+     *
      * @param jigyoshaHenkoJiyu 事業者変更事由
      */
     public void setJigyoshaHenkoJiyu(RString jigyoshaHenkoJiyu) {
@@ -280,9 +281,9 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * このエンティティの主キーが他の{@literal DbT3006KyotakuKeikakuJigyoshaSakuseiEntity}と等しいか判定します。
-     * 
+     *
      * @param other 比較するエンティティ
-     * @return 
+     * @return
      * 比較するエンティティが同じ主キーを持つ{@literal DbT3006KyotakuKeikakuJigyoshaSakuseiEntity}の場合{@literal true}、それ以外の場合は{@literal false}
      */
     @Override
@@ -321,6 +322,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
 
     /**
      * {@inheritDoc}
+     *
      * @return {@inheritDoc}
      */
     @Override
@@ -329,6 +331,4 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiEntity extends DbTableEntityBas
     }
 
 // </editor-fold>
-
-
 }
